@@ -1,58 +1,69 @@
-var tape = require("tape"),
-    d3 = require("../"),
-    d3_geo = require("d3-geo");
-require("./projectionEqual");
+import {
+  geoAirocean,
+  geoCahillKeyes,
+  geoComplexLog,
+  geoCubic,
+  geoDodecahedral,
+  geoIcosahedral,
+  geoImago,
+  geoImagoBlock,
+  geoTetrahedralLee,
+} from "../src/index.js";
+import { assertProjectionEqual } from "./asserts.js";
 
-
-tape("inverse polyhedrals", function(test) {
-
-  [ d3.geoAirocean(), d3.geoCubic(), d3.geoIcosahedral(), d3.geoDodecahedral() ]
-  .forEach(function(projection) {
-    [ [-23, 12], [10,10], [100,-45] ]
-    .forEach(function(location) {
-      projection.angle(Math.random()*360);
-      test.projectionEqual(projection, location, projection(location), 1e-5);
-    });
-  });
-  test.end();
+it("inverse polyhedrals", function () {
+  [geoAirocean(), geoCubic(), geoIcosahedral(), geoDodecahedral()].forEach(
+    function (projection) {
+      [
+        [-23, 12],
+        [10, 10],
+        [100, -45],
+      ].forEach(function (location) {
+        projection.angle(Math.random() * 360);
+        assertProjectionEqual(projection, location, projection(location), 1e-5);
+      });
+    }
+  );
 });
 
-tape("inverse Imago, tetrahedralLee", function(test) {
-
-  [ d3.geoImago(), d3.geoImagoBlock(), d3.geoTetrahedralLee() ]
-  .forEach(function(projection) {
-    [ [-23, 12], [10,10], [100,-45] ]
-    .forEach(function(location) {
-      projection.angle(Math.random()*360);
-      test.projectionEqual(projection, location, projection(location), 1e-5);
+it("inverse Imago, tetrahedralLee", function () {
+  [geoImago(), geoImagoBlock(), geoTetrahedralLee()].forEach(function (
+    projection
+  ) {
+    [
+      [-23, 12],
+      [10, 10],
+      [100, -45],
+    ].forEach(function (location) {
+      projection.angle(Math.random() * 360);
+      assertProjectionEqual(projection, location, projection(location), 1e-5);
     });
   });
-  test.end();
 });
 
-tape("inverse Cahill-Keyes", function(test) {
-
-  [ d3.geoCahillKeyes() ]
-  .forEach(function(projection) {
-    [ [-23, 12], [10,10], [100,-45] ]
-    .forEach(function(location) {
-      projection.angle(Math.random()*360);
-      test.projectionEqual(projection, location, projection(location), 1e-5);
+it("inverse Cahill-Keyes", function () {
+  [geoCahillKeyes()].forEach(function (projection) {
+    [
+      [-23, 12],
+      [10, 10],
+      [100, -45],
+    ].forEach(function (location) {
+      projection.angle(Math.random() * 360);
+      assertProjectionEqual(projection, location, projection(location), 1e-5);
     });
   });
-  test.end();
 });
 
-tape("inverse complex log", function(test) {
-
-  [ d3.geoComplexLog() ]
-  .forEach(function(projection) {
-    [ [0, 0], [-23, 12], [10,10], [100,-45] ]
-    .forEach(function(location) {
-      projection.angle(Math.random()*360);
-      test.projectionEqual(projection, location, projection(location), 1e-5);
+it("inverse complex log", function () {
+  [geoComplexLog()].forEach(function (projection) {
+    [
+      [0, 0],
+      [-23, 12],
+      [10, 10],
+      [100, -45],
+    ].forEach(function (location) {
+      projection.angle(Math.random() * 360);
+      assertProjectionEqual(projection, location, projection(location), 1e-5);
     });
   });
-  test.end();
 });
-
