@@ -1,19 +1,34 @@
-import {terser} from "rollup-plugin-terser";
-import * as meta from "./package.json";
+import { readFileSync } from "fs";
+import terser from "@rollup/plugin-terser";
+import meta from "./package.json" assert { type: "json" };
+
+// Extract copyrights from the LICENSE.
+const copyright = readFileSync("./LICENSE", "utf-8")
+  .split(/\n/g)
+  .filter((line) => /^Copyright\s+/.test(line))
+  .map((line) => line.replace(/^Copyright\s+/, ""))
+  .join(", ");
 
 const config = {
   input: "src/index.js",
-  external: Object.keys(meta.dependencies || {}).filter(key => /^d3-/.test(key)),
+  external: Object.keys(meta.dependencies || {}).filter((key) =>
+    /^d3-/.test(key)
+  ),
   output: {
     file: `dist/${meta.name}.js`,
     name: "d3",
     format: "umd",
     indent: false,
     extend: true,
-    banner: `// ${meta.homepage} v${meta.version} Copyright ${(new Date).getFullYear()} ${meta.author.name}`,
-    globals: Object.assign({}, ...Object.keys(meta.dependencies || {}).filter(key => /^d3-/.test(key)).map(key => ({[key]: "d3"})))
+    banner: `// ${meta.homepage} v${meta.version} Copyright ${copyright}`,
+    globals: Object.assign(
+      {},
+      ...Object.keys(meta.dependencies || {})
+        .filter((key) => /^d3-/.test(key))
+        .map((key) => ({ [key]: "d3" }))
+    ),
   },
-  plugins: []
+  plugins: [],
 };
 
 export default [
@@ -22,15 +37,15 @@ export default [
     ...config,
     output: {
       ...config.output,
-      file: `dist/${meta.name}.min.js`
+      file: `dist/${meta.name}.min.js`,
     },
     plugins: [
       ...config.plugins,
       terser({
         output: {
-          preamble: config.output.banner
-        }
-      })
-    ]
-  }
+          preamble: config.output.banner,
+        },
+      }),
+    ],
+  },
 ];
